Fakeit example in PyXspec
This Fakeit example in PyXspec shows how to trick XSPEC into producing a type-II file that holds 10,000 spectra for the 'n3' detector.  Because the n3 detector is originally of type-I format, Fakeit defaults output to a type-I format file.  You need to trick Fakeit with a type-II file (the 'lle' file).  We do there here.
*** This Notebook names the file with model name "grbm+pow" and that does NOT allow for a type-II output file for some reason.  If you change the filename from "fakespec_n3_grbm+pow_-01-_L_.fak" to "fakespec_n3_grbm-pow_-01-_L_.fak", it will output a type-II file as desired.

Spectra files are provided as well as the out files.